Add files via upload
Heavenly Father, I'm so grateful for your willingness to come into my life and help me through every storm. I'm thankful for your strength and power, and I trust that you will never let me down. When storms come, I know I can always turn to you for help. So today, I'm asking you to help me find peace in my family, job, career, and education.And he arose, and rebuked the wind, and said unto the sea, Peace, be still. And the wind ceased, and there was a great calm.As Christians, we have the power of faith to calm any storm that comes our way. Jesus showed us this by believing in his own power and commanding the storm to be still. We don't have to be afraid when life throws us a curveball, because Jesus is with us and has the power to stop it. When the storm came, the disciples of Jesus cried out to him for help.